Update Field Formatting
This fixes the issue where a number of spaces were inserted before fields. This was actually the number of parameters/2.

This ensures there is always 1 space, and adds "[" "]" around fields to separate them from the main content.

This means

WithFields("test", true, "Error", "not so serious").Warnln("This is your last.")

will log as "WARN   This is your last. [test=true, Error=not so serious]"